Add navigation to the 'BlogPostScreen'
Adds a new screen, 'BlogPostScreen', and integrates it into the app's navigation flow. The `HealthBlogPage` now allows navigation to `BlogPostScreen` when clicking the image in the carousel and there is also navigation to the Home and Welcome Screens in the `HealthBlogPage`. Also adds a new route for the `BlogPostScreen` in `Navigation.kt`.